Separate type equality checker from contract update validator
